api/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "A simple MCP client"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "bs4>=0.0.2",
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "langchain-mcp-adapters>=0.1.4",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "loguru>=0.7.3",
    "mcp[cli]>=1.9.2",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
